Minor cleanup of CFamily helper code
* extracted path calculations to a separate class
* moved standalone execution method from CFamilyHelper to CLangAnalyzer
